gratitude_django/requirements.txt
certifi==2018.8.13
dj-database-url==0.5.0
Django==2.0.5
django-cors-headers==2.4.0
django-heroku==0.3.1
django-tastypie==0.14.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
psycopg2==2.7.5
PyJWT==1.6.4
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.5
six==1.11.0
whitenoise==4.0
